ID Data type Long 으로 변경 및 쿼리문 작성

+++ src/main/java/dance/withgnu/demo/user/repository/VideoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface VideoRepository extends JpaRepository<Video, Long> {
-    List<Video> findByUserId(int userId);
+    List<Video> findByUserId(Long userId);
 }
